bcachefs: fix a dio write bug

+++ fs/bcachefs/io_types.h
@@ -94,7 +94,7 @@ struct bch_write_op {
 
 	unsigned		written; /* sectors */
 	u16			flags;
-	s8			error;
+	s16			error; /* dio write path expects it to hold -ERESTARTSYS... */
 
 	unsigned		csum_type:4;
 	unsigned		compression_type:4;
